same commit
I just dont know what stashed changes are, now I restored them, we will see which is the latesr

+++ include/pid.h
@@ -3,6 +3,6 @@
 #ifndef PID_H
 #define PID_H
 
-int pid(int rpm, int setpoint);
+float pid(int rpm, int setpoint);
 
 #endif
